some work on animation

+++ resources/engine_code/shaders/raycast.cs.glsl
@@ -22,8 +22,8 @@ uniform layout(rgba8) image3D lighting;
 uniform int x_offset;
 uniform int y_offset;
 
-uniform int clickndragx;
-uniform int clickndragy;
+uniform float clickndragx;
+uniform float clickndragy;
 
 //gl_GlobalInvocationID will define the tile size, so doing anything to define it here would be redundant
 // this shader is general up to tile sizes of 2048x2048, since those are the maximum dispatch values
@@ -133,7 +133,7 @@ vec4 get_color_for_pixel(vec3 org, vec3 dir)
 
 void main()
 {
-    ivec2 Global_Loc = ivec2(gl_GlobalInvocationID.xy) + ivec2(x_offset+clickndragx, y_offset+clickndragy);
+    ivec2 Global_Loc = ivec2(gl_GlobalInvocationID.xy) + ivec2(x_offset+int(clickndragx), y_offset+int(clickndragy));
     ivec2 dimensions = ivec2(imageSize(current));
 
     float aspect_ratio = float(dimensions.y) / float(dimensions.x);
